feat(retend-server): Add misc metadata support
This commit introduces support for a `misc` property in the `PageMeta` type. This allows developers to include custom metadata that is not directly related to standard HTML meta tags. The `meta.js` file has been updated to ignore the `misc` property during standard meta tag updates. The `client.js` file has been updated to include the `misc` property in the `PageMeta` type definition.