Refactor: Rename fragment creation methods
This commit renames `createFragment` to `createNodeGroup` and
`unwrapFragment` to `unwrapNodeGroup`. This is a more accurate
description of what these methods do, as they can create and unwrap
groups of nodes, not just fragments.

Additionally, `isFragment` has been renamed to `isGroup` to match the
new naming convention. These changes improve the clarity and
maintainability of the codebase.